변경내역 DL3 duration -> 시:분:초 포맷으로 변경 파일 크기에 대한 계산법 변경(1024) DL3 상세조회조건 추가 (단말정보) 인터넷 익스플로어 관련 처리 변경 테이블 크기 약간 조정(세로길이)

+++ MAMBrowser/BLL/PrivateFileBLL.cs
@@ -115,11 +115,14 @@ namespace MAMBrowser.BLL
             if (seqList.Count <= 0)
                 return;
 
-            //파일 실제 삭제
+            //물리파일 삭제
             foreach (var seq in seqList)
             {
                 var fileData = Get(seq);
-                _fileProtocol.Delete(fileData.FilePath);
+                if (File.Exists(fileData.FilePath))
+                {
+                    _fileProtocol.Delete(fileData.FilePath);
+                }
             }
 
             _dao.Delete(userId, seqList);
